Add mockito to the project and mock the class MyPoint

+++ MyPoint.java
@@ -223,8 +223,8 @@ public class MyPoint implements ITranslation {
 	 * @param random2 The random number generator used for y.
 	 */
 	public void setPoint(final Random random1, final Random random2) {
-		setX(random1.nextInt());
-		setY(random2.nextInt());
+		setX(random1.nextDouble());
+		setY(random2.nextDouble());
 	}
